sonar-duplications/src/main/java/org/sonar/duplications/statement/matcher/ExactTokenMatcher.java
/*
 * SonarQube
 * Copyright (C) 2009-2019 SonarSource SA
 * mailto:info AT sonarsource DOT com
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with this program; if not, write to the Free Software Foundation,
 * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
package org.sonar.duplications.statement.matcher;

import java.util.List;

import org.sonar.duplications.token.Token;
import org.sonar.duplications.token.TokenQueue;

/**
 * Consumes only one specified token.
 */
public class ExactTokenMatcher extends TokenMatcher {

  private final String tokenToMatch;

  public ExactTokenMatcher(String tokenToMatch) {
    if (tokenToMatch == null) {
      throw new IllegalArgumentException();
    }
    this.tokenToMatch = tokenToMatch;
  }

  @Override
  public boolean matchToken(TokenQueue tokenQueue, List<Token> matchedTokenList) {
    if (tokenQueue.isNextTokenValue(tokenToMatch)) {
      matchedTokenList.add(tokenQueue.poll());
      return true;
    }
    return false;
  }

}
